Move unit tests to their respective modules
