+++ platform-sdk/consensus-model/build.gradle.kts
@@ -8,6 +8,16 @@ plugins {
 
 description = "Consensus Model"
 
+testModuleInfo {
+    requires("com.swirlds.platform.core")
+    requires("com.swirlds.platform.core.test.fixtures")
+    requires("org.hiero.base.utility.test.fixtures")
+    requires("org.hiero.consensus.model.test.fixtures")
+    requires("org.hiero.consensus.utility.test.fixtures")
+    requires("org.junit.jupiter.api")
+    requires("org.mockito")
+}
+
 timingSensitiveModuleInfo {
     requires("com.swirlds.common.test.fixtures")
     requires("org.hiero.base.concurrent")
