+++ platform-sdk/consensus-model/src/test/java/org/hiero/consensus/model/event/CesEventTest.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.event;
+package org.hiero.consensus.model.event;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,8 +7,6 @@ import java.io.IOException;
 import org.hiero.base.constructable.ConstructableRegistryException;
 import org.hiero.base.utility.test.fixtures.io.InputOutputStream;
 import org.hiero.consensus.constructable.ConstructableRegistration;
-import org.hiero.consensus.model.event.CesEvent;
-import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.test.fixtures.event.TestingEventBuilder;
 import org.hiero.consensus.test.fixtures.Randotron;
 import org.junit.jupiter.api.BeforeAll;
