+++ platform-sdk/consensus-model/src/test/java/org/hiero/consensus/model/event/PlatformEventTest.java
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.event;
+package org.hiero.consensus.model.event;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.swirlds.platform.event.EventSerializationUtils;
 import com.swirlds.platform.test.fixtures.utils.EqualsVerifier;
 import java.io.IOException;
 import java.util.Random;
 import org.hiero.base.utility.test.fixtures.RandomUtils;
-import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.test.fixtures.event.TestingEventBuilder;
 import org.hiero.consensus.test.fixtures.Randotron;
 import org.junit.jupiter.api.DisplayName;
