+++ platform-sdk/consensus-model/src/test/java/org/hiero/consensus/model/transaction/TransactionWrapperTest.java
@@ -1,10 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.system.transaction;
+package org.hiero.consensus.model.transaction;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.hedera.pbj.runtime.io.buffer.Bytes;
-import org.hiero.consensus.model.transaction.TransactionWrapper;
 import org.hiero.consensus.test.fixtures.Randotron;
 import org.junit.jupiter.api.Test;
